生单流程 完成 50%

+++ taki-mall-order/src/main/java/com/taki/order/domin/request/CreateOrderRequest.java
@@ -206,7 +206,7 @@ public class CreateOrderRequest extends AbstractObject implements Serializable {
         /**
          * 费用
          */
-        private Integer amount;
+        private BigDecimal amount;
     }
 
     /**
